refactor(types): menubar
